refactor!: Restructure package to accomodate both RPC and REST clients
