+++ docs/deprecations.md
@@ -1,0 +1,8 @@
+# Deprecation timeline
+
+## v1.0.0
+
+* `citric.Client` will be renamed to `citric.RPC`.
+* `citric.client` will be moved to `citric.rpc.client`.
+* `citric.session` will be moved to `citric.rpc.session`.
+* `citric.method` will be moved to `citric.rpc.method`.
