+++ src/citric/rpc/__init__.py
@@ -1,0 +1,7 @@
+"""RPC client for LimeSurvey."""
+
+from __future__ import annotations
+
+from .client import RPC
+
+__all__ = ["RPC"]
